cli: Migrate initial database directly to MongoDB
Adds the command 'db-migrate' that migrates the initial database and creates
other database structures that are necessary for the application. This
process communicates directly with MongoDB. The old migration process that
uses HTTP is left for stress testing the powerplant server.

The migration process is faster and more robust now, it should be easier to
set up a working environment from scratch.