client: Removed unused code
Remove all code in client/ that is unused and won't be directly used for the
next release. Some of it may be partly reintroduced in different form though.
Hopefully this makes it easier to navigate the code.

Currently the app consists of just the offline-editable crops list. The logic
for searching and selecting compatible crops for beds needs to be rewritten.